+++ src/noctilux/__init__.py
@@ -2,4 +2,4 @@
 
 __all__ = ["__version__"]
 
-__version__ = "0.7.4"
+__version__ = "0.7.5"